reorganized the directories for easier expansion. added comments for easier expansion.
