+++ src/main/java/com/geneticselection/GeneticSelection.java
@@ -1,21 +1,15 @@
 package com.geneticselection;
 
-import com.geneticselection.custommobs.Cows.CustomCowEntity;
+import com.geneticselection.mobs.Cows.CustomCowEntity;
+import com.geneticselection.mobs.ModEntities;
 import net.fabricmc.api.ModInitializer;
 
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
 import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.minecraft.block.Blocks;
 import net.minecraft.entity.*;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.Heightmap;
-import net.minecraft.world.World;
-import net.minecraft.world.WorldAccess;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.biome.BiomeKeys;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -23,41 +17,34 @@ public class GeneticSelection implements ModInitializer {
 	public static final String MOD_ID = "genetic-selection";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
-	private ActionResult onEntitySpawn(Entity entity, World world, double x, double y, double z) {
-		// Check if the entity is a vanilla cow
-		if (entity.getType() == EntityType.COW) {
-			// Cancel the cow spawn event
-			return ActionResult.FAIL;
-		}
-		// Allow other entities to spawn normally
-		return ActionResult.PASS;
-	}
-
 	public void cowMethod(){
+		//Register the default attibutes to your mob
 		FabricDefaultAttributeRegistry.register(ModEntities.CUSTOM_COW, CustomCowEntity.createCowAttributes());
-
+		//lowers the spawn rate of default vanilla cows
 		BiomeModifications.addSpawn(
-				BiomeSelectors.foundInOverworld(),
-				SpawnGroup.CREATURE,
-				EntityType.COW, // Remove the original cow
-				0, // 0 value means removal
-				0, // No minimum group size
-				0  // No maximum group size
+			BiomeSelectors.foundInOverworld(),
+			SpawnGroup.CREATURE,
+			EntityType.COW, // Remove the original cow
+			0, // Spawn weight (higher = more frequent)
+			0, // No minimum group size
+			0  // No maximum group size
 		);
+		//adds custom cow to natural spawn
 		BiomeModifications.addSpawn(
-				BiomeSelectors.foundInOverworld(),
-				SpawnGroup.CREATURE,
-				ModEntities.CUSTOM_COW, // Add custom cow
-				20, // Spawn weight (higher = more frequent)
-				2,  // Minimum group size
-				4   // Maximum group size
+			BiomeSelectors.foundInOverworld(),
+			SpawnGroup.CREATURE,
+			ModEntities.CUSTOM_COW, // Add custom cow
+			20, // Spawn weight (higher = more frequent)
+			2,  // Minimum group size
+			4   // Maximum group size
 		);
+		//restricts the cow spawn to grass blocks on the ground
 		SpawnRestriction.register(
-				ModEntities.CUSTOM_COW,
-				SpawnLocationTypes.ON_GROUND,
-				Heightmap.Type.MOTION_BLOCKING_NO_LEAVES,
-				(entityType, world, spawnReason, pos, random) ->
-						world.getBlockState(pos.down()).isOf(Blocks.GRASS_BLOCK)
+			ModEntities.CUSTOM_COW,
+			SpawnLocationTypes.ON_GROUND,
+			Heightmap.Type.MOTION_BLOCKING_NO_LEAVES,
+			(entityType, world, spawnReason, pos, random) ->
+				world.getBlockState(pos.down()).isOf(Blocks.GRASS_BLOCK)
 		);
 	}
