+++ src/main/java/com/geneticselection/mobs/ModEntities.java
@@ -1,6 +1,7 @@
-package com.geneticselection;
+package com.geneticselection.mobs;
 
-import com.geneticselection.custommobs.Cows.CustomCowEntity;
+import com.geneticselection.GeneticSelection;
+import com.geneticselection.mobs.Cows.CustomCowEntity;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.minecraft.entity.EntityDimensions;
 import net.minecraft.entity.EntityType;
@@ -17,4 +18,5 @@ public class ModEntities{
                     .dimensions(EntityDimensions.fixed(0.9F, 1.4F))
                     .build()
     );
+    //register new mobs here by using the above format
 }
