+++ src/main/java/com/geneticselection/mobs/ModModleLayers.java
@@ -1,0 +1,10 @@
+package com.geneticselection.mobs;
+
+import com.geneticselection.GeneticSelection;
+import net.minecraft.client.render.entity.model.EntityModelLayer;
+import net.minecraft.util.Identifier;
+
+public class ModModleLayers {
+    public static final EntityModelLayer CUSTOM_COW = new EntityModelLayer(Identifier.of(GeneticSelection.MOD_ID,"custom_cow"), "main");
+    //add your new mobs here by using the above format
+}
